requirements.txt
numpy==1.26.0
pandas==2.1.1
scipy==1.11.1
matplotlib==3.7.1
statsmodels==0.14.0
celery==5.3.4
psycopg2-binary==2.9.7
redis==4.5.5
pyyaml==6.0.1 